requirements.txt
streamlit
pandas
gspread
gspread-dataframe
google-auth
google-auth-oauthlib
google-auth-httplib2